backend/requirements.txt
# Django Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Machine Learning & Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Database
# SQLite is included with Python

# Additional utilities (if needed)
python-dateutil>=2.8.0
